Computed ppty eg
If you uncomment the functions in method, and comment computed property, you will see that the methods are executed always regardless of whether the value of the bound variable is changed or not. Computed properties are cached based on their dependencies. A computed property will only re-evaluate when some of its dependencies have changed. This means as long as isHealthy/age has not changed, multiple access to the addAge & insuranceDecider computed property will immediately return the previously computed result without having to run the function again.